fix(maven): use direct stream output for execution listener logging
SLF4J loggers in the ClassRealm don't connect to batch-runner output.
Pass output stream through invocation chain and write directly to it.
